refactor the agent
